Extract shared variable-label helper into idata_utils
