+++ pymc_extras/inference/laplace_approx/idata.py
@@ -1,4 +1,3 @@
-from itertools import product
 from typing import Literal
 
 import numpy as np
@@ -14,52 +13,7 @@ from scipy.optimize import OptimizeResult
 from scipy.sparse.linalg import LinearOperator
 from xarray import DataTree
 
-
-def make_default_labels(name: str, shape: tuple[int, ...]) -> list:
-    if len(shape) == 0:
-        return [name]
-
-    return [list(range(dim)) for dim in shape]
-
-
-def make_unpacked_variable_names(
-    names: list[str], model: pm.Model, var_name_to_model_var: dict[str, str] | None = None
-) -> list[str]:
-    coords = model.coords
-    initial_point = model.initial_point()
-
-    if var_name_to_model_var is None:
-        var_name_to_model_var = {}
-
-    value_to_dim = {
-        value.name: model.named_vars_to_dims.get(model.values_to_rvs[value].name, None)
-        for value in model.value_vars
-    }
-    value_to_dim = {k: v for k, v in value_to_dim.items() if v is not None}
-
-    rv_to_dim = model.named_vars_to_dims
-    dims_dict = rv_to_dim | value_to_dim
-
-    unpacked_variable_names = []
-    for name in names:
-        name = var_name_to_model_var.get(name, name)
-        shape = initial_point[name].shape
-        if shape:
-            dims = dims_dict.get(name)
-            if dims:
-                labels_by_dim = [
-                    coords[dim] if shape[i] == len(coords[dim]) else np.arange(shape[i])
-                    for i, dim in enumerate(dims)
-                ]
-            else:
-                labels_by_dim = make_default_labels(name, shape)
-            labels = product(*labels_by_dim)
-            unpacked_variable_names.extend(
-                [f"{name}[{','.join(map(str, label))}]" for label in labels]
-            )
-        else:
-            unpacked_variable_names.extend([name])
-    return unpacked_variable_names
+from pymc_extras.inference.idata_utils import make_unpacked_variable_names
 
 
 def map_results_to_inference_data(
